Refactor : 서버쪽에서 socket 이벤트 등록시 game과 signaling관련 기능을 분리해서 등록
상세 :
- service디렉토리 아래에 game과 signaling디렉토리를 생성 후 각각의
  디렉토리에서 이벤트를 가져와 등록하도록 구현
- sendCandidate 이벤트명을 sendIceCandidate로 변경
이유 :
- 다른 기능을 가진 이벤트들을 분리하여 등록함으로써 유지보수성을 증가시키기 위함

close #330
