+++ client/src/service/StreamingManager.js
@@ -15,7 +15,7 @@ class StreamingManager {
   registerSocketEvents() {
     this.socket.on('assignStreamer', this.assignStreamerHandler.bind(this));
     this.socket.on('assignViewer', this.assignViewerHandler.bind(this));
-    this.socket.on('sendCandidate', this.sendCandidateHandler.bind(this));
+    this.socket.on('sendIceCandidate', this.sendIceCandidateHandler.bind(this));
     this.socket.on('sendDescription', this.sendDescriptionHandler.bind(this));
   }
 
@@ -85,8 +85,8 @@ class StreamingManager {
     webRTCManager.registerTrack(streamerSocketId, trackHandler.bind(this));
   }
 
-  async sendCandidateHandler({ target, candidate }) {
-    await this.webRTCManager.addIceCandidate(target, candidate);
+  async sendIceCandidateHandler({ target, iceCandidate }) {
+    await this.webRTCManager.addIceCandidate(target, iceCandidate);
   }
 
   async sendDescriptionHandler({ target, description }) {
@@ -100,9 +100,9 @@ class StreamingManager {
 
   iceCandidateHandler(socketId, event) {
     if (!event.candidate) return;
-    this.socket.emit('sendCandidate', {
+    this.socket.emit('sendIceCandidate', {
       target: socketId,
-      candidate: event.candidate,
+      iceCandidate: event.candidate,
     });
   }
