+++ server/app.js
@@ -2,7 +2,7 @@ const express = require('express');
 const path = require('path');
 const cookieParser = require('cookie-parser');
 const logger = require('morgan');
-const gameService = require('./gameService');
+const service = require('./service');
 
 const app = express();
 
@@ -22,6 +22,6 @@ app.use((req, res) => {
   res.redirect('/');
 });
 
-app.gameService = gameService;
+app.service = service;
 
 module.exports = app;
