+++ server/service/game/eventHandlers/index.js
@@ -1,8 +1,6 @@
 const matchHandler = require('./matchHandler');
 const sendReadyHandler = require('./sendReadyHandler');
 const sendChattingMessageHandler = require('./sendChattingMessageHandler');
-const sendCandidateHandler = require('./sendCandidateHandler');
-const sendDescriptionHandler = require('./sendDescriptionHandler');
 const askSocketIdHandler = require('./askSocketIdHandler');
 // const disconnectingHandler = require('./disconnectingHandler');
 
@@ -10,8 +8,6 @@ module.exports = {
   matchHandler,
   sendReadyHandler,
   sendChattingMessageHandler,
-  sendCandidateHandler,
-  sendDescriptionHandler,
   askSocketIdHandler,
   // disconnectingHandler,
 };
