+++ server/service/game/eventHandlers/sendChattingMessageHandler.js
@@ -1,5 +1,5 @@
 const short = require('short-uuid');
-const { io } = require('../io');
+const { io } = require('../../io');
 const roomController = require('../roomController');
 
 const sendChattingMessageHandler = (socket, { nickname, message }) => {
