+++ server/service/signaling/eventHandlers/index.js
@@ -1,0 +1,7 @@
+const sendIceCandidateHandler = require('./sendIceCandidateHandler');
+const sendDescriptionHandler = require('./sendDescriptionHandler');
+
+module.exports = {
+  sendIceCandidateHandler,
+  sendDescriptionHandler,
+};
