+++ server/service/signaling/eventHandlers/sendDescriptionHandler.js
@@ -1,4 +1,4 @@
-const { io } = require('../io');
+const { io } = require('../../io');
 
 const sendDescriptionHandler = (socket, { target, description }) => {
   io.to(target).emit('sendDescription', { target: socket.id, description });
