+++ server/service/signaling/eventHandlers/sendIceCandidateHandler.js
@@ -1,0 +1,7 @@
+const { io } = require('../../io');
+
+const sendIceCandidateHandler = (socket, { target, iceCandidate }) => {
+  io.to(target).emit('sendIceCandidate', { target: socket.id, iceCandidate });
+};
+
+module.exports = sendIceCandidateHandler;
